commit ../patches/anon-core.patch
It allows you to have all the SBCL heap in anonymous VM pages, as
opposed to having the image backed pages be backed by the image file
(with copy-on-write) and then the rest of the heap anonymously.

This is not currently used.  Can be used for e.g. experimenting with
hugepages.  A major drawback is that nothing is shared between
different Lisp instances started from the same image anymore.