Commit ../patches/before-gc-hook.patch
I/we insist on having a before-gc hook like CMUCL and SBCL used to.

It was removed from SBCL due to being unsafe since you might/will cons
while already trying to GC.  We did not notice any problems with
this.  We mostly use it for bookkeeping and separating out performance
measuraments so that GC isn't mixed in with mainline performance data.

I/we definitely want this.