commit ../patches/nonexecutable-stack.patch
Mark the ELF object as not requiring an executable stack.

This shuts up various complaining tools that look at executable stacks
for security reasons.

This is properly ifdefed for Linux and only needs testing on non-x86
to be committed.

Although I don't quite understand how this works with stack-allocated
closures...